Implement reconnecting and handle chat messages
The server restarts everyday at 4 a.m so the bot needs to reconnect
if connection is lost. Moreover there are sometimes restarts of just
one mode and the client may be kicked to the hub so there has to be
a 'rejoin from hub' system. All these actions ought to be done with
random delay due to the fact that there are usually more bots per one
computer and the server is not happy when they try to connect or join
a mode at the same time.

- Log in and connect to the specified mode after connected to the server
and try again with short random delay if not succeeded
- Rejoin the mode if kicked to the hub
- Listen to Keep Alive packets, respond if received or disconnect if not
received for 20 seconds
- Handle chat messages, change them into text objects with styles
and pure readable text